backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
pymongo
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
openai-whisper
Pillow
python-docx
pdfplumber
requests
beautifulsoup4